new edit to also know the sequence
now know the sequences and the frequency and the Hash value related to them. due to hashing retreival should be fast and I have also now ensured that the actual sequence can be remembered as well though it would be faster to only keep hash value and frequency I decided to play it safe.